fix: 規約違反の修正 (命名規則、型エイリアス、Doxygen、ドキュメント)
- 命名規則の修正 (PascalCase化):
  - prefix_sum_2d -> PrefixSum2d
  - power_table -> PowerTable
  - rolling_hash -> RollingHash
  - xorshift -> XorShift

- 型エイリアスの適用 (gwen/types.hpp):
  - size_t -> usize, int -> i32, unsigned long long -> u64
  - 対象: prefix_sum, bit_vector, wavelet_matrix, rolling_hash,
    idxsort, tree_lca, grid, constants, matrix, compress, modint

- index.md へのドキュメント登録修正:
  - types.md, dump.md, monoid.md, abel.md, fenwick_tree.md の
    リンクを正しいパスに修正
  - wavelet_matrix.md のパスを container -> ds に修正
  - lis.md を新規追加

- Doxygenコメントの追加:
  - abel.hpp, monoid.hpp, mod.hpp, template.hpp

- ドキュメントのAC Libraryスタイル適用:
  - idxsort.md, runlength.md, edge.md, grid.md, dijkstra.md,
    macro.md, constants.md, prefix_sum.md, dump.md, bit.md,
    utils.md, graph.md

- テスト: 105/105 全合格

+++ include/gwen/mod/modint.hpp
@@ -15,7 +15,7 @@ namespace gwen {
  * ACLのmodintと互換性のあるインターフェースを持つことを要求する。
  */
 template <typename T>
-concept modint = ring<T> && requires(T a, T b, unsigned long long n) {
+concept modint = ring<T> && requires(T a, T b, u64 n) {
     { a / b } -> std::same_as<T>;
     { a /= b } -> std::same_as<T&>;
     { a.val() } -> std::integral;
@@ -55,7 +55,7 @@ public:
         assert(n_ & 1);
         n = n_;
         ns = n;
-        for (int i = 0; i < 5; ++i) ns *= 2 - ns * n;
+        for (i32 i = 0; i < 5; ++i) ns *= 2 - ns * n;
         assert(ns * n == 1);
         ns = -ns;
         r2 = -static_cast<u128>(n) % n;
